Add files via upload
Customer Analytics in Python is where marketing and data science meet. Data science and marketing are two of the key driving forces that help companies create value and stay on top in today’s fast-paced economy. My project is packed with knowledge, and includes sections on customer and purchase analytics, as well as a deep-learning model, all implemented in Python.